fix: change the location of common helper methods
